Code refactoring and reorganization.
