+++ lib/pages/hotels_and_restaurants.dart
@@ -3,12 +3,12 @@ import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:lottie/lottie.dart';
 
+import '../localization/supported_languages.dart';
+import '../navigation/bottom_navigation.dart';
 import '../navigation/cutom_app_bar.dart';
 import '../navigation/navigation_helper.dart';
-import '../navigation/bottom_navigation.dart';
-import '../localization/supported_languages.dart';
-import 'hotels_and_restaurants/hotels/hotels.dart';
-import 'hotels_and_restaurants/restaurants/restaurants.dart';
+import 'hotels/hotels.dart';
+import 'restaurants/restaurants.dart';
 
 class HotelsAndRestaurants extends StatefulWidget {
   const HotelsAndRestaurants({Key? key}) : super(key: key);
